Jbide 21857: Hot class reload doesn't work on OpenShift: incorporate pull request comments (part 1)
