+++ plugins/org.jboss.tools.jmx.remoting1_1_3/.classpath
@@ -1,16 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-logging-3.1.4.GA.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-logging-processor-1.1.0.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-logmanager-1.5.3.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-marshalling-1.4.8.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-marshalling-river-1.4.8.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-remoting-3.3.3.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jboss-sasl-1.0.4.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jdeparser-1.0.0.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/remoting-jmx-1.1.3.Final.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/xnio-api-3.0.10.GA.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/xnio-nio-3.0.10.GA.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
 		<attributes>
 			<attribute name="maven.pomderived" value="true"/>
@@ -23,10 +12,16 @@
 			<attribute name="maven.pomderived" value="true"/>
 		</attributes>
 	</classpathentry>
-	<classpathentry kind="con" path="org.eclipse.m2e.MAVEN2_CLASSPATH_CONTAINER">
-		<attributes>
-			<attribute name="maven.pomderived" value="true"/>
-		</attributes>
-	</classpathentry>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-logging-3.1.4.GA.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-logging-processor-1.1.0.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-logmanager-1.5.3.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-marshalling-1.4.8.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-marshalling-river-1.4.8.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-remoting-3.3.3.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jboss-sasl-1.0.4.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jdeparser-1.0.0.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/remoting-jmx-1.1.3.Final.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/xnio-api-3.0.10.GA.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/xnio-nio-3.0.10.GA.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
