+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/core/jmx/OpenShiftServerAdapter.java
@@ -1,3 +1,13 @@
+/*******************************************************************************
+ * Copyright (c) 2016 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v1.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
 package org.jboss.tools.openshift.core.jmx;
 
 import org.eclipse.core.runtime.IAdapterFactory;
@@ -5,6 +15,11 @@ import org.jboss.tools.jmx.core.ExtensionManager;
 import org.jboss.tools.jmx.core.IConnectionWrapper;
 import org.jboss.tools.openshift.core.server.OpenShiftServer;
 
+/**
+ * Adapt OpenshiftServer to a IConnectionWrapper
+ * @author Thomas Mäder
+ *
+ */
 public class OpenShiftServerAdapter implements IAdapterFactory {
 	private static final Class<?>[] ADAPTERS = new Class[]{IConnectionWrapper.class};
