+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/core/jmx/VersionDetector.java
@@ -1,3 +1,13 @@
+/*******************************************************************************
+ * Copyright (c) 2016 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v1.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
 package org.jboss.tools.openshift.core.jmx;
 
 import java.util.HashMap;
@@ -14,6 +24,11 @@ import com.openshift.restclient.model.IBuildConfig;
 import com.openshift.restclient.model.IProject;
 import com.openshift.restclient.model.IService;
 
+/**
+ * Detects the version of EAP/Wildfly running as an OpenShift server.
+ * @author Thomas Mäder
+ *
+ */
 public class VersionDetector {
 	private static final Map<String, VersionKey> knowImages= buildVersionMap();
 	
@@ -28,6 +43,10 @@ public class VersionDetector {
 		return result;
 	}
 
+	/**
+	 * Guess the version key
+	 * @return null if no matching version is found
+	 */
 	VersionKey guess() {
 		IService service = OpenShiftServerUtils.getService(server);
 		if (service != null) {
@@ -36,8 +55,12 @@ public class VersionDetector {
 				List<IBuildConfig> buildConfigs = project.getResources(ResourceKind.BUILD_CONFIG);
 				if (buildConfigs != null) {
 					IBuildConfig buildConfig = ResourceUtils.getBuildConfigForService(service, buildConfigs);
-					DockerImageURI imageUri = OpenShiftServerUtils.getImageUri(buildConfig);
-					return knowImages.get(imageUri.getName());
+					if (buildConfig != null) {
+						DockerImageURI imageUri = ResourceUtils.getImageUri(buildConfig);
+						if (imageUri != null) {
+							return knowImages.get(imageUri.getName());
+						}
+					}
 				}
 			}
 		}
