+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/core/jmx/VersionKey.java
@@ -1,5 +1,20 @@
+/*******************************************************************************
+ * Copyright (c) 2016 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v1.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
 package org.jboss.tools.openshift.core.jmx;
 
+/**
+ * Identifies a particular version of a web server
+ * @author Thomas Mäder
+ *
+ */
 public final class VersionKey {
 	public VersionKey(ServerType serverType, String version) {
 		this.serverType = serverType;
